revert changes and add upgrade search/uninstall
