+++ src/AppInstallerCLICore/AppInstallerCLICore.vcxproj.filters
@@ -176,7 +176,7 @@
     <ClInclude Include="Public\COMContext.h">
       <Filter>Public</Filter>
     </ClInclude>
-    <ClInclude Include="Workflows\PortableInstallFlow.h">
+    <ClInclude Include="Workflows\PortableFlow.h">
       <Filter>Workflows</Filter>
     </ClInclude>
   </ItemGroup>
@@ -322,7 +322,7 @@
     <ClCompile Include="Workflows\DownloadFlow.cpp">
       <Filter>Workflows</Filter>
     </ClCompile>
-    <ClCompile Include="Workflows\PortableInstallFlow.cpp">
+    <ClCompile Include="Workflows\PortableFlow.cpp">
       <Filter>Workflows</Filter>
     </ClCompile>
   </ItemGroup>
