+++ src/AppInstallerCLICore/Workflows/UninstallFlow.h
@@ -30,12 +30,6 @@ namespace AppInstaller::CLI::Workflow
     // Outputs: None
     void MsixUninstall(Execution::Context& context);
 
-    // Removes the Portable package.
-    // Required Args: None
-    // Inputs: ProductCode
-    // Outputs: None
-    void PortableUninstall(Execution::Context& context);
-
     // Records the uninstall to the tracking catalog.
     // Required Args: None
     // Inputs: Package
